<chore>[sdk]: update SSO sdk
Related: ZSV-7198
