--- sdk/src/main/java/org/zstack/sdk/CasState.java
@@ -1,6 +1,0 @@
-package org.zstack.sdk;
-
-public enum CasState {
-	Enabled,
-	Disabled,
-}
