+++ sdk/src/main/java/org/zstack/sdk/OAuth2ClientInventory.java
@@ -2,7 +2,7 @@ package org.zstack.sdk;
 
 
 
-public class OAuth2ClientInventory extends org.zstack.sdk.SSOClientInventory {
+public class OAuth2ClientInventory extends org.zstack.sdk.identity.imports.entity.ThirdPartyAccountSourceInventory {
 
     public java.lang.String clientId;
     public void setClientId(java.lang.String clientId) {
@@ -12,12 +12,28 @@ public class OAuth2ClientInventory extends org.zstack.sdk.SSOClientInventory {
         return this.clientId;
     }
 
-    public java.lang.String clientSecret;
-    public void setClientSecret(java.lang.String clientSecret) {
-        this.clientSecret = clientSecret;
+    public java.lang.String grantType;
+    public void setGrantType(java.lang.String grantType) {
+        this.grantType = grantType;
     }
-    public java.lang.String getClientSecret() {
-        return this.clientSecret;
+    public java.lang.String getGrantType() {
+        return this.grantType;
+    }
+
+    public java.lang.String loginMNUrl;
+    public void setLoginMNUrl(java.lang.String loginMNUrl) {
+        this.loginMNUrl = loginMNUrl;
+    }
+    public java.lang.String getLoginMNUrl() {
+        return this.loginMNUrl;
+    }
+
+    public java.lang.String redirectUrl;
+    public void setRedirectUrl(java.lang.String redirectUrl) {
+        this.redirectUrl = redirectUrl;
+    }
+    public java.lang.String getRedirectUrl() {
+        return this.redirectUrl;
     }
 
     public java.lang.String authorizationUrl;
@@ -44,20 +60,20 @@ public class OAuth2ClientInventory extends org.zstack.sdk.SSOClientInventory {
         return this.userinfoUrl;
     }
 
-    public java.lang.String grantType;
-    public void setGrantType(java.lang.String grantType) {
-        this.grantType = grantType;
-    }
-    public java.lang.String getGrantType() {
-        return this.grantType;
-    }
-
     public java.lang.String logoutUrl;
     public void setLogoutUrl(java.lang.String logoutUrl) {
         this.logoutUrl = logoutUrl;
     }
     public java.lang.String getLogoutUrl() {
         return this.logoutUrl;
+    }
+
+    public java.lang.String usernameProperty;
+    public void setUsernameProperty(java.lang.String usernameProperty) {
+        this.usernameProperty = usernameProperty;
+    }
+    public java.lang.String getUsernameProperty() {
+        return this.usernameProperty;
     }
 
 }
