+++ sdk/src/main/java/org/zstack/sdk/UpdateOAuthClientAction.java
@@ -53,10 +53,10 @@ public class UpdateOAuthClientAction extends AbstractAction {
     public java.lang.String userinfoUrl;
 
     @Param(required = false, nonempty = false, nullElements = false, emptyString = true, noTrim = false)
-    public java.lang.String loginType;
+    public java.lang.String logoutUrl;
 
     @Param(required = false, nonempty = false, nullElements = false, emptyString = true, noTrim = false)
-    public java.lang.String logoutUrl;
+    public java.lang.String usernameProperty;
 
     @Param(required = false)
     public java.util.List systemTags;
