+++ sdk/src/main/java/org/zstack/sdk/identity/ldap/entity/LdapServerInventory.java
@@ -2,39 +2,7 @@ package org.zstack.sdk.identity.ldap.entity;
 
 
 
-public class LdapServerInventory  {
-
-    public java.lang.String uuid;
-    public void setUuid(java.lang.String uuid) {
-        this.uuid = uuid;
-    }
-    public java.lang.String getUuid() {
-        return this.uuid;
-    }
-
-    public java.lang.String name;
-    public void setName(java.lang.String name) {
-        this.name = name;
-    }
-    public java.lang.String getName() {
-        return this.name;
-    }
-
-    public java.lang.String type;
-    public void setType(java.lang.String type) {
-        this.type = type;
-    }
-    public java.lang.String getType() {
-        return this.type;
-    }
-
-    public java.lang.String description;
-    public void setDescription(java.lang.String description) {
-        this.description = description;
-    }
-    public java.lang.String getDescription() {
-        return this.description;
-    }
+public class LdapServerInventory extends org.zstack.sdk.identity.imports.entity.ThirdPartyAccountSourceInventory {
 
     public java.lang.String url;
     public void setUrl(java.lang.String url) {
@@ -76,22 +44,6 @@ public class LdapServerInventory  {
         return this.encryption;
     }
 
-    public java.sql.Timestamp createDate;
-    public void setCreateDate(java.sql.Timestamp createDate) {
-        this.createDate = createDate;
-    }
-    public java.sql.Timestamp getCreateDate() {
-        return this.createDate;
-    }
-
-    public java.sql.Timestamp lastOpDate;
-    public void setLastOpDate(java.sql.Timestamp lastOpDate) {
-        this.lastOpDate = lastOpDate;
-    }
-    public java.sql.Timestamp getLastOpDate() {
-        return this.lastOpDate;
-    }
-
     public java.lang.String filter;
     public void setFilter(java.lang.String filter) {
         this.filter = filter;
@@ -106,22 +58,6 @@ public class LdapServerInventory  {
     }
     public java.lang.String getUsernameProperty() {
         return this.usernameProperty;
-    }
-
-    public java.lang.String createAccountStrategy;
-    public void setCreateAccountStrategy(java.lang.String createAccountStrategy) {
-        this.createAccountStrategy = createAccountStrategy;
-    }
-    public java.lang.String getCreateAccountStrategy() {
-        return this.createAccountStrategy;
-    }
-
-    public java.lang.String deleteAccountStrategy;
-    public void setDeleteAccountStrategy(java.lang.String deleteAccountStrategy) {
-        this.deleteAccountStrategy = deleteAccountStrategy;
-    }
-    public java.lang.String getDeleteAccountStrategy() {
-        return this.deleteAccountStrategy;
     }
 
 }
